Blog/requirements.txt
distlib==0.3.6
django-crispy-forms==1.14.0
filelock==3.8.0
Pillow==9.3.0
platformdirs==2.5.4
virtualenv==20.16.7
